requirements.txt
python-chess==1.999
numpy
matplotlib
moviepy  # if you plan to create videos programmatically
